add logic user can only have access, update or delete his order

+++ src/apps/order/order/api/order.controller.ts
@@ -1,6 +1,7 @@
 import { Request, Response, NextFunction } from "express";
 import { OrderService } from "../domain/order.service";
 import { ApiResponse } from "../../../../librairies/controllers/api.response";
+import BadRequestError from "../../../../config/error/bad.request.config";
 
 export class OrderController {
   private orderService: OrderService;
@@ -20,7 +21,8 @@ export class OrderController {
 
   async getAllOrders(req: Request, res: Response, next: NextFunction): Promise<void> {
     try {
-      const orders = await this.orderService.getAllOrders();
+      const userId = req.user.id;
+      const orders = await this.orderService.getOrdersByUserId(userId);
       ApiResponse.success(res, "Orders retrieved successfully", orders, 200);
     } catch (error) {
       next(error);
@@ -29,7 +31,15 @@ export class OrderController {
 
   async getOrderById(req: Request, res: Response, next: NextFunction): Promise<void> {
     try {
-      const order = await this.orderService.getOrderById(req.params.id);
+      const orderId = req.params.id;
+      const userId = req.user.id;
+      const order = await this.orderService.getOrderById(orderId);
+      if (!order || order.user.toString() !== userId) {
+        throw new BadRequestError({
+          message: "Order does not belong to the authenticated user.",
+          logging: true,
+        });
+      }
       ApiResponse.success(res, "Order retrieved successfully", order, 200);
     } catch (error) {
       next(error);
@@ -40,6 +50,14 @@ export class OrderController {
     try {
       const orderId = req.params.id;
       const updates = req.body;
+      const userId = req.user.id;
+      const order = await this.orderService.getOrderById(orderId);
+      if (!order || order.user.toString() !== userId) {
+        throw new BadRequestError({
+          message: "Order does not belong to the authenticated user.",
+          logging: true,
+        });
+      }
       const updatedOrder = await this.orderService.updateOrderById(orderId, updates);
       ApiResponse.success(res, "Order updated successfully", updatedOrder, 200);
     } catch (error) {
@@ -50,6 +68,14 @@ export class OrderController {
   async deleteOrder(req: Request, res: Response, next: NextFunction): Promise<void> {
     try {
       const orderId = req.params.id;
+      const userId = req.user.id;
+      const order = await this.orderService.getOrderById(orderId);
+      if (!order || order.user.toString() !== userId) {
+        throw new BadRequestError({
+          message: "Order does not belong to the authenticated user.",
+          logging: true,
+        });
+      }
       await this.orderService.deleteOrderById(orderId);
       ApiResponse.success(res, "Order deleted successfully", null, 204);
     } catch (error) {
